Commit the transaction in dev or prod mode only
Since we have multiple threads running some of the threads which
are listening to Kafka Messages when they create an object in the
DB the main thread doesn't see it unless a COMMIT is issued.
This also causes challenges when testing since tests use the Rails
Transaction/Rollback to create temporary records which get removed
when the test runs. If we force a commit the objects stay in the DB
and affect the tests. For now I have have made the commit only for
development and production mode.